rename plugins to commands and update readme
