+++ README.md
@@ -1,6 +1,6 @@
 ## ![A lion with a crown and the text lion](https://i.imgur.com/AX8My57.png)
 
-Lion is a **plugin-oriented** [Discord](https://discordapp.com/) bot that utilizes Discord.js and TypeScript. Lion was created originally to satisfy the needs of server automation within the [UCF CS Discord](https://discord.gg/KXdcdxZPTV).
+Lion is a **command-oriented** [Discord](https://discordapp.com/) bot that utilizes Discord.js and TypeScript. Lion was created originally to satisfy the needs of server automation within the [UCF CS Discord](https://discord.gg/KXdcdxZPTV).
 
 # Getting Started
 
@@ -20,38 +20,29 @@ Execute the following command to launch Lion:
 
 All subsequent saved changes made within the `/src` directory will make Lion automatically rebuild while the `watch` command is running.
 
-## Your first plugin
+## Your first command
 
-Navigate to the `/src/app/plugins` directory and create a new file following the specific schema:
-`example.plugin.ts`
+Navigate to the `/src/app/commands` directory and create a new file following the specific schema:
+`example.command.ts`
 
 For ease of use, it's recommended that your plugin name is in all lowercase and in snake_case.
 
-**Your plugin must be the _default_ export of the file: e.g. `export default class ExamplePlugin`**
+**Your plugin must be the _default_ export of the file: e.g. `export default { ... } satisfies Command`**
 
 Here's a skeleton for how a basic plugin should look:
 
 ```typescript
-import { Plugin } from '../../common/plugin';
-import { IContainer, IMessage, ChannelType } from '../../common/types';
+import { Command } from '../../common/slash';
 
-export default class ExamplePlugin extends Plugin {
-  public commandName: string = 'example';
-  public name: string = 'Example Plugin';
-  public description: string = 'An example plugin to show how awesome Lion is!';
-  public usage: string = 'example';
-  public permission: ChannelType = ChannelType.Bot;
-
-  constructor(public container: IContainer) {
-    super();
-  }
-
-  public async execute(message: IMessage, args?: string[]) {
+export default {
+  name: 'example',
+  description: 'An example command to show how awesome Lion is!',
+  async execute({ interaction, container }) {
     // Your logic here.
-  }
-}
+  },
+} satisfies Command;
 ```
 
-Again, this is a basic skeleton. It is suggested to read the abstract `Plugin` class to fully know what is possible within a plugin, which is located in `/src/common/plugin.ts`.
+Again, this is a basic skeleton. It is suggested to read the `Command` type to fully know what is possible within a command, which is located in `/src/common/slash.ts`.
 
-And that's it! You have successfully created your first plugin with Lion!
+And that's it! You have successfully created your first command with Lion!
