+++ src/app/commands/alt.command.ts
