+++ src/app/commands/anonreport.command.ts
@@ -1,6 +1,6 @@
 import { Command } from '../../common/slash';
 
-const plugin = {
+const command = {
   commandName: 'anonreport',
   name: 'anonreport',
   description: 'anonymously report a concern to moderation team',
@@ -33,4 +33,4 @@ const plugin = {
   },
 } satisfies Command;
 
-export default plugin;
+export default command;
