+++ src/app/commands/cointoss.command.ts
