+++ src/app/commands/connectfour.command.ts
