+++ src/app/commands/diceroll.command.ts
