+++ src/app/commands/equation.command.ts
