+++ src/app/commands/listclasses.command.ts
