+++ src/app/commands/poll.command.ts
