+++ src/app/commands/register.command.ts
