+++ src/app/commands/ta.command.ts
