+++ src/app/commands/unregister.command.ts
